+++ src/test/Common.hs
@@ -15,8 +15,6 @@ module Common
   , solvedWith
   , solvedWithout
   , solvedUsing
-  , getGas
-  , gasInRange
   , countCorpus
   , overrideQuiet
   , loadSolTests
@@ -48,7 +46,6 @@ import Echidna.Config (parseConfig, defaultConfig)
 import Echidna.Campaign (runWorker)
 import Echidna.Solidity (selectMainContract, mkTests, loadSpecified, compileContracts)
 import Echidna.Test (checkETest)
-import Echidna.Types (Gas)
 import Echidna.Types.Config (Env(..), EConfig(..), EConfigWithUsage(..))
 import Echidna.Types.Campaign
 import Echidna.Types.Signature (ContractName)
@@ -247,15 +244,6 @@ solvedWith tx t final =
 solvedWithout :: TxCall -> Text -> (Env, WorkerState) -> IO Bool
 solvedWithout tx t final =
   maybe False (all $ (/= tx) . (.call)) <$> solnFor t final
-
-getGas :: Text -> WorkerState -> Maybe (Gas, [Tx])
-getGas t camp = Map.lookup t camp.gasInfo
-
-gasInRange :: Text -> Gas -> Gas -> (Env, WorkerState) -> IO Bool
-gasInRange t l h (_, workerState) = do
-  pure $ case getGas t workerState of
-    Just (g, _) -> g >= l && g <= h
-    _           -> False
 
 countCorpus :: Int -> (Env, WorkerState) -> IO Bool
 countCorpus n (env, _) = do